rename python to ips_python
- need to have an unambiguous python module so that we can do recognizable imports, instead of relative imports, which have been causing issues
