+++ ips_python/cosine.py
@@ -3,12 +3,12 @@ import pickle
 import pandas as pd
 from sklearn.metrics.pairwise import cosine_similarity
 
-from .utils import get_data_path
+from ips_python.utils import get_data_path
 from os.path import join
 
-from .preprocessing import preprocess_query_text
-from .vectorize import vectorize_input_text
-from .constants import (
+from ips_python.preprocessing import preprocess_query_text
+from ips_python.vectorize import vectorize_input_text
+from ips_python.constants import (
     PROCESSED_RECORDS_FILENAME,
     TERM_DOCUMENT_MATRIX_FILENAME,
     VECTORIZER_FILENAME,
